+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## API
 * add "environmental" of symbols in /meta's response
+* support simple `display` theory, with `goal`, `relevant`, `expanded` and `view`
+  * replace `__goals(Term)` by `__relevant(Term)`
 
 
 ## [0.5.1] - 2020-05-26